code/aleksandr-k3s-istio2/requirements.txt
Flask==2.2.5
gunicorn
requests==2.31.0
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-requests

